refactored database submodule from file to folder
